+++ src/main/java/ch/admin/bj/swiyu/didtoolbox/model/CryptographicAlgorithm.java
@@ -2,7 +2,7 @@ package ch.admin.bj.swiyu.didtoolbox.model;
 
 public enum CryptographicAlgorithm {
     ED25519("Ed25519"),
-    ECP256("EC P-256");
+    ECP256("EcP-256");
 
     private final String name;